requirements.txt
BeautifulReport==0.1.3
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
idna==3.4
parameterized==0.9.0
PyYAML==6.0
requests==2.29.0
urllib3==1.26.15
